Poprawa błędów z konsultacji

+++ role_manager.c
@@ -1,8 +1,9 @@
 #include "common.h"
 
-// Ścieżka do programu obsługi (wymagane przez exec)
+// Ścieżka do programu obsługi klienta
 #define SERVICE_EXEC_PATH "./service"
 
+// Przygotowanie procesu potomnego (stanowiska dynamicznego)
 void prepare_dynamic_child()
 {
     signal(SIGTSTP, SIG_IGN);
@@ -10,45 +11,50 @@ void prepare_dynamic_child()
     signal(SIGCONT, SIG_DFL);
 }
 
+// Flaga kontynuacji pracy menedżera
 volatile int keep_running = 1;
+
+// Handler sygnału SIGTERM
 void mgr_sigterm(int sig)
 {
     (void)sig;
     keep_running = 0;
 }
 
+// Główna funkcja menedżera - zarządza stanowiskami obsługi i mechaników
 void run_manager(int shm_id, int sem_id, int msg_id)
 {
     global_sem_id = sem_id;
     shared_data *shm = (shared_data *)safe_shmat(shm_id, NULL, 0);
     
-    // Zapamiętanie start_time
+    // Kopiowanie czasu startu symulacji
     time_t start_time = shm->start_time;
 
-    // Przygotowanie argumentów (stringi) dla exec
+    // Przygotowanie argumentów dla exec (konwersja int -> string)
     char msg_id_str[16], shm_id_str[16], sem_id_str[16];
     snprintf(msg_id_str, sizeof(msg_id_str), "%d", msg_id);
     snprintf(shm_id_str, sizeof(shm_id_str), "%d", shm_id);
     snprintf(sem_id_str, sizeof(sem_id_str), "%d", sem_id);
 
-    // Ustawienie handlera SIGTERM
+    // Rejestracja handlera SIGTERM
     signal(SIGTERM, mgr_sigterm);
     log_color(ROLE_MANAGER, "MANAGER: Start.");
 
-    // Pętla zarządzania
+    // Bufor na komendy sterujące z klawiatury
     char buffer[100];
     fd_set readfds;
     struct timeval tv;
 
+    // Główna pętla zarządzania
     while (keep_running && shm->simulation_running)
     {
-        // Sprawdzenie wejścia standardowego (sterowanie ręczne)
+        // Sprawdzanie wejścia z klawiatury (sterowanie ręczne)
         FD_ZERO(&readfds);
         FD_SET(STDIN_FILENO, &readfds);
         tv.tv_sec = 1;
         tv.tv_usec = 0;
 
-        // Czytanie wejścia
+        // Nieblokujące czytanie z klawiatury
         int ret = select(STDIN_FILENO + 1, &readfds, NULL, NULL, &tv);
         if (ret > 0 && FD_ISSET(STDIN_FILENO, &readfds))
         {
@@ -57,8 +63,10 @@ void run_manager(int shm_id, int sem_id, int msg_id)
             {
                 buffer[bytes] = 0;
                 int sig_num, mech_id;
+                // Parsowanie komendy: numer_sygnału numer_mechanika
                 if (sscanf(buffer, "%d %d", &sig_num, &mech_id) >= 1)
                 {
+                    // Komenda 4 - pożar (zamknięcie całego serwisu)
                     if (sig_num == 4)
                     {
                         log_color(ROLE_MANAGER, "MANAGER: POŻAR! Zamykam serwis.");
@@ -67,36 +75,34 @@ void run_manager(int shm_id, int sem_id, int msg_id)
                                 kill(shm->mechanics_pids[i], SIG_FIRE);
                         shm->simulation_running = 0;
                     }
+                    // Komendy 1-3 dla konkretnego mechanika
                     else if (mech_id >= 1 && mech_id <= NUM_MECHANICS)
                     {
+                        // Ignorowanie zamknięcia już zamkniętego stanowiska
                         if (sig_num == 1 && shm->mechanic_status[mech_id - 1] == 2)
                             continue;
                         pid_t target = shm->mechanics_pids[mech_id - 1];
                         if (target > 0)
                         {
                             int s = 0;
-                            if (sig_num == 1)
-                                s = SIG_CLOSE_STATION;
-                            if (sig_num == 2)
-                                s = SIG_SPEED_UP;
-                            if (sig_num == 3)
-                                s = SIG_RESTORE_SPEED;
-                            if (s)
-                                kill(target, s);
+                            if (sig_num == 1) s = SIG_CLOSE_STATION;  // Zamknięcie stanowiska
+                            if (sig_num == 2) s = SIG_SPEED_UP;       // Przyspieszenie
+                            if (sig_num == 3) s = SIG_RESTORE_SPEED;  // Normalna prędkość
+                            if (s) kill(target, s);
                         }
                     }
                 }
             }
         }
 
-        // Obliczenie aktualnego czasu symulacji
+        // Obliczanie aktualnego czasu symulacji
         sim_time st = get_simulation_time(start_time);
         
         // Zarządzanie stanowiskami obsługi na podstawie długości kolejki
         int q = shm->clients_in_queue;
         int is_open_time = (st.hour >= OPEN_HOUR && st.hour < CLOSE_HOUR);
 
-        // Stanowisko 2
+        // Stanowisko 2 - uruchamiane gdy kolejka > K1
         if (shm->service_pids[1] == 0 && q > K1 && is_open_time)
         {
             log_color(ROLE_MANAGER, "MANAGER: Kolejka %d > %d. Uruchamiam Obsługę 2.", q, K1);
@@ -112,11 +118,13 @@ void run_manager(int shm_id, int sem_id, int msg_id)
         }
         else if (shm->service_pids[1] > 0)
         {
+            // Sprawdzanie czy stanowisko 2 zakończyło pracę
             int status;
             if (waitpid(shm->service_pids[1], &status, WNOHANG) > 0)
             {
                 shm->service_pids[1] = 0;
             }
+            // Zamykanie stanowiska 2 gdy kolejka <= 2 lub poza godzinami
             else if (q <= 2 || !is_open_time)
             {
                 kill(shm->service_pids[1], SIGTERM);
@@ -125,7 +133,7 @@ void run_manager(int shm_id, int sem_id, int msg_id)
             }
         }
 
-        // Stanowisko 3
+        // Stanowisko 3 - uruchamiane gdy kolejka > K2
         if (shm->service_pids[2] == 0 && q > K2 && is_open_time)
         {
             log_color(ROLE_MANAGER, "MANAGER: Kolejka %d > %d. Uruchamiam Obsługę 3.", q, K2);
@@ -141,11 +149,13 @@ void run_manager(int shm_id, int sem_id, int msg_id)
         }
         else if (shm->service_pids[2] > 0)
         {
+            // Sprawdzanie czy stanowisko 3 zakończyło pracę
             int status;
             if (waitpid(shm->service_pids[2], &status, WNOHANG) > 0)
             {
                 shm->service_pids[2] = 0;
             }
+            // Zamykanie stanowiska 3 gdy kolejka <= 3 lub poza godzinami
             else if (q <= 3 || !is_open_time)
             {
                 kill(shm->service_pids[2], SIGTERM);
@@ -154,7 +164,7 @@ void run_manager(int shm_id, int sem_id, int msg_id)
             }
         }
 
-        // Sprawdzenie czy wszyscy mechanicy zamknięci
+        // Sprawdzanie czy wszystkie stanowiska mechaników zamknięte
         int all_mech_closed = 1;
         for (int i = 0; i < NUM_MECHANICS; i++)
         {
@@ -165,7 +175,7 @@ void run_manager(int shm_id, int sem_id, int msg_id)
             }
         }
 
-        // Zamknięcie symulacji jeśli wszyscy mechanicy zamknięci
+        // Zamknięcie symulacji gdy wszyscy mechanicy zamknięci
         if (all_mech_closed)
         {
             log_color(ROLE_MANAGER, "MANAGER: Wszystkie stanowiska mechaników zamknięte. Zamykam system.");
@@ -173,7 +183,7 @@ void run_manager(int shm_id, int sem_id, int msg_id)
         }
     }
 
-    // Zamykanie procesów obsługi
+    // Zamykanie dynamicznych stanowisk obsługi
     for (int i = 1; i < NUM_SERVICE_STAFF; i++)
     {
         if (shm->service_pids[i] > 0)
@@ -183,15 +193,17 @@ void run_manager(int shm_id, int sem_id, int msg_id)
         }
     }
 
-    // Flush kolejki klientów
+    // Opróżnianie kolejki klientów przy zamknięciu
     msg_buf msg;
     while (msgrcv(msg_id, &msg, sizeof(msg_buf) - sizeof(long), MSG_CLIENT_TO_SERVICE, IPC_NOWAIT) != -1)
     {
-        // Zwalniamy slot w kolejce i dekrementujemy licznik
+        // Zwalnianie slotu w kolejce
         sem_lock(sem_id, SEM_LOG);
-        shm->clients_in_queue--;
+        if (shm->clients_in_queue > 0)
+            shm->clients_in_queue--;
         sem_unlock(sem_id, SEM_LOG);
         sem_unlock(sem_id, SEM_QUEUE);
+        // Wysyłanie odmowy do klienta
         msg.mtype = MSG_BASE_TO_CLIENT + msg.client_pid;
         msg.decision = 0;
         msgsnd(msg_id, &msg, sizeof(msg_buf) - sizeof(long), 0);
